# WSO2 Enterprise Integration Eclipse project/packageServices/packageServicesConfigs/src/main/resources/metadata/Packages_swagger.yaml
openapi: 3.0.1
info:
  title: Packages
  description: API Definition of Packages
  version: 1.0.0
servers:
- url: http://localhost:8290/packages
paths:
  /getpackages:
    get:
      responses:
        default:
          description: Default response
  /addpackage:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response
  /getpackagebyid:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response
  /editpackage:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response
  /deletepackage:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response
  /adddd:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response
  /editbyid:
    post:
      requestBody:
        description: Sample Payload
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
        required: false
      responses:
        default:
          description: Default response